chore: rename content dir
