+++ contents/enablement-page/resources-framework.yml
