+++ contents/enablement-page/resources-manual.yml
