+++ contents/enablement-page/resources.yml
