+++ contents/strategy-page/index.yml
