+++ contents/strategy-page/resources.yml
