+++ gatsby-config.js
@@ -13,8 +13,8 @@ module.exports = {
     {
       resolve: 'gatsby-source-filesystem',
       options: {
-        name: 'content',
-        path: `${__dirname}/content`,
+        name: 'contents',
+        path: `${__dirname}/contents`,
       },
     },
     {
