+++ gatsby-node.js
@@ -7,7 +7,7 @@ exports.onCreateNode = ({ node, getNode, actions }) => {
   // Ensures we are processing only markdown files
   if (node.internal.type === 'MarkdownRemark') {
     const parent = getNode(node.parent);
-    if (parent.sourceInstanceName === 'content') {
+    if (parent.sourceInstanceName === 'contents') {
       createNodeField({
         node,
         name: 'slug',
