+++ src/components/home/Benefit.js
@@ -7,8 +7,8 @@ import { HeroGraphic } from '../hero/HeroGraphic.js';
 import { Hero } from '../hero/Hero.js';
 import { AccordionEntry } from '../AccordionEntry.js';
 import { useAccordion } from '../../hooks/useAccordion.js';
-import heroContent from '../../../content/home-page/index.yml';
-import benefits from '../../../content/home-page/benefits.yml';
+import heroContent from '../../../contents/home-page/index.yml';
+import benefits from '../../../contents/home-page/benefits.yml';
 
 const {
   heroBenefit: { category, title, body },
