+++ src/components/home/HowToContribute.js
@@ -3,8 +3,8 @@ import { Card, CardBody, Icon } from 'design-react-kit';
 import { MobileSwiper } from '../MobileSwiper.js';
 import { Hero } from '../hero/Hero.js';
 import { ExternalLink } from '../ExternalLink.js';
-import heroContent from '../../../content/home-page/index.yml';
-import contributeCards from '../../../content/home-page/contribute.yml';
+import heroContent from '../../../contents/home-page/index.yml';
+import contributeCards from '../../../contents/home-page/contribute.yml';
 
 const {
   heroContribute: { category, title, body },
