+++ src/pages/programma-abilitazione-cloud.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { StaticImage } from 'gatsby-plugin-image';
 import { createUseStyles } from 'react-jss';
 import { graphql, useStaticQuery } from 'gatsby';
-import content from '../../content/enablement-page/index.yml';
-import resources from '../../content/enablement-page/resources.yml';
-import resourcesManual from '../../content/enablement-page/resources-manual.yml';
-import resourcesFramework from '../../content/enablement-page/resources-framework.yml';
+import content from '../../contents/enablement-page/index.yml';
+import resources from '../../contents/enablement-page/resources.yml';
+import resourcesManual from '../../contents/enablement-page/resources-manual.yml';
+import resourcesFramework from '../../contents/enablement-page/resources-framework.yml';
 import { HeroTitle } from '../components/hero/HeroTitle.js';
 import { HeroBody } from '../components/hero/HeroBody.js';
 import { Hero } from '../components/hero/Hero.js';
